Region isn't deleted if it has a flower

+++ controllers/region/delete_post.js
@@ -1,9 +1,23 @@
 const asyncHandler = require('express-async-handler')
 const path = require('path')
 const Region = require(path.join(appRoot, 'models', 'region.js'))
+const Flower = require(path.join(appRoot, 'models', 'flower.js'))
 
 module.exports = asyncHandler(async (req, res, next) => {
-  await Region.findByIdAndDelete(req.body.region)
+  const regionId = req.body.region
 
-  res.status(200).send("Region successfully deleted!")
+  if (! await regionHasFlower(regionId)) {
+    await Region.findByIdAndDelete(req.body.region)
+    res.status(200).send("Region successfully deleted!")
+  } else {
+    res.status(400).send("Region has a flower")
+  }
 })
+
+async function regionHasFlower(regionId) {
+  const flower = await Flower.findOne({ region: regionId}).exec()
+  if (flower === null)
+    return false
+  else
+    return true
+}
